requirements.txt
mysql_connector_repackaged==0.3.1
pycryptodome==3.21.0
